Complete Version 1.1
Finished the base game - you are a character represented by an asterik ( * ), going through a "maze" of vertical walls that have gaps in them. You can go either up or down within the "columns", and can go on to the next "column" once you are at where the gap in the wall is. There is an instant-win cheat code and unlisted commands within the code, which are more or less unimportant. If you run into a wall (including the top and bottom walls), you die.